Update decorator
Fix Linters
