+++ decorator.rb
@@ -1,13 +1,30 @@
-require './nameable'
+class Nameable
+  def correct_name
+    raise NotImplementedError, "#{self.class} has not implemented method #{__method__}"
+  end
+end
 
 class Decorator < Nameable
   attr_accessor :nameable
 
   def initialize(nameable)
     @nameable = nameable
+    super()
   end
 
   def correct_name
     @nameable.correct_name
+  end
+end
+
+class CapitalizeDecorator < Decorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
+
+class TrimmerDecorator < Decorator
+  def correct_name
+    @nameable.correct_name.slice(0, 9)
   end
 end
