+++ person.rb
@@ -1,4 +1,4 @@
-require './nameable'
+require './decorator'
 
 class Person < Nameable
   def initialize(age, name = 'Unknown', parent_permission: true)
@@ -6,6 +6,7 @@ class Person < Nameable
     @name = name
     @age = age
     @parent_permission = parent_permission
+    super()
   end
 
   def correct_name
